% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appServer.R
\name{appServer}
\alias{appServer}
\title{Server Logic for shinyDashboard}
\usage{
appServer(input, output, session)
}
\arguments{
\item{input, output, session}{Standard Shiny server arguments.}
}
\description{
Server Logic for shinyDashboard
}
\keyword{internal}
