% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_longitudinal_table.R
\name{assessment_longitudinal_table}
\alias{assessment_longitudinal_table}
\title{Assessment Longitudinal Table}
\usage{
assessment_longitudinal_table(
  dat,
  id = "NACCID",
  date = "VISITDATE",
  descriptions = c(Impaired = 0.03, Borderline = 0.1, `Low Average` = 0.26, Average =
    0.76, `High Average` = 0.92, Superior = 0.97, `Very Superior` = 1),
  fill_values,
  methods = list(),
  table_font_size = 100,
  table_id = NULL,
  show_all_visits = TRUE,
  stubhead_label = NULL
)
}
\arguments{
\item{dat}{A data.table object.}

\item{id}{A character string specifying a column in the data frame \code{dat}. Default is \code{"NACCID"}.}

\item{date}{A character string specifying a column in the data frame \code{dat}. Default is \code{"VISITDATE"}.}

\item{descriptions}{A named numeric vector. Default has names \code{"Impaired"}, \code{"Borderline"}, \code{"Low Average"}, \code{"Average"}, \code{"High Average"}, \code{"Superior"}, and \code{"Very Superior"} with corresponding values \code{0.03}, \code{0.10}, \code{0.26}, \code{0.76}, \code{0.92}, \code{0.97}, and \code{1}.}

\item{fill_values}{Optional.}

\item{methods}{Either a list containing methods used for standardization (each a character vector with named entried \code{method} and \code{version}), or \code{"infer"} (default). If \code{"infer"}, then methods are pulled from attributes of standardized columns.}

\item{table_font_size}{A numeric value passed to \code{gt::tab_options(table.font.size = gt::pct(table_font_size))}. Defaults to 100.}

\item{table_id}{Optional. ID given to the table. If not provided, random string assigned.}

\item{show_all_visits}{A boolean. If \code{TRUE} (default), all visits present in the data are shown in the table. If \code{FALSE}, only visits with at least one standardized score are presented.}

\item{stubhead_label}{Optional. Passed to \code{gt::tab_stubhead(label = stubhead_label)}.}
}
\value{
A \code{shiny::HTML} table created from a \code{gt::gt} object.
}
\description{
Create a table of raw scores across multiple visits colored by the corresponding standardized scores.
}
