% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_summary_table.R
\name{assessment_summary_table}
\alias{assessment_summary_table}
\alias{assessment_summary_data}
\title{Create NACC T-Cog Neuropsychological Assessment Summary Table}
\usage{
assessment_summary_table(summary_dat, bar_height = 16)

assessment_summary_data(
  dat,
  id = "NACCID",
  descriptions = c(Impaired = 0.03, Borderline = 0.1, `Low Average` = 0.26, Average =
    0.76, `High Average` = 0.92, Superior = 0.97, `Very Superior` = 1),
  fill_values = NULL,
  include_caption = TRUE
)
}
\arguments{
\item{summary_dat}{A list as returned by \code{assessment_summary_data()}.}

\item{bar_height}{In pixels. Height of the percentile bars. Default: 16}

\item{dat}{Data to use. Must have exactly one row (data from one participant and one visit).}

\item{id}{Column containing ID.}

\item{descriptions}{A named vector giving cutoffs to use for creating descriptions.}

\item{fill_values}{(optional) A named vector of same length (with same names)
as \code{descriptions} with hex color values to use. If \code{NULL}, no colorcoding used.
By default, evenly spread out colors from red through yellow to green are used.}

\item{include_caption}{Logical; should a caption be included above the table
with age and IQCODEs at current visit?}
}
\value{
An object of class \code{gt::gt_tbl}
}
\description{
This function create a summary table of measures from the NACC T-Cog Neuropsychological
Assessment.
}
\keyword{internal}
