% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_plots.R
\name{base_plot_z_scores}
\alias{base_plot_z_scores}
\title{Create base plotly figure}
\usage{
base_plot_z_scores(
  x_range,
  y_range,
  descriptions = c(Impaired = 0.03, Borderline = 0.1, `Low Average` = 0.26, Average =
    0.76, `High Average` = 0.92, Superior = 0.97, `Very Superior` = 1),
  fill_values,
  shade_descriptions = T,
  fill_alpha = 0.2,
  source = "A",
  new_id = NULL
)
}
\arguments{
\item{x_range}{vector with two elements giving the smallest and largest values to use on x-axis (in order)}

\item{y_range}{vector with two elements giving the smallest and largest values to use on y-axis (in order)}

\item{descriptions}{A named vector giving cutoffs to use for creating descriptions.}

\item{fill_values}{(optional) A named vector of same length (with same names)
as \code{descriptions} with hex color values to use. If \code{NULL}, no colorcoding used.
By default, evenly spread out colors from red through yellow to green are used.}

\item{shade_descriptions}{logical; should ranges for descriptions be shaded?}

\item{fill_alpha}{opacity value for shaded areas}

\item{source}{passed to \code{plotly::plot_ly}}

\item{new_id}{Optional. String to use as the plotly source ID. If \code{NULL}
(default), a random string is assigned.}
}
\description{
Create a base plot where traces can later be added.
}
\keyword{internal}
