% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainTableModule.R
\name{chrome_extra_args}
\alias{chrome_extra_args}
\title{Return Chrome CLI arguments}
\usage{
chrome_extra_args(default_args = c("--disable-gpu"))
}
\arguments{
\item{default_args}{Arguments to be used in any circumstances.}
}
\value{
A character vector with CLI arguments to be passed to Chrome.
}
\description{
This is a helper function which returns arguments to be passed to Chrome.
This function tests whether the code is running on shinyapps and returns the
appropriate Chrome extra arguments. Source: https://github.com/RLesur/chrome_print_shiny
}
\keyword{internal}
