% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colSelectModule.R, R/methodSelectModule.R
\name{colSelectUI}
\alias{colSelectUI}
\alias{colSelectServer}
\alias{colSelectApp}
\alias{methodSelectUI}
\alias{methodSelectServer}
\alias{methodSelectApp}
\title{Shiny Module for Selecting Columns}
\usage{
colSelectUI(id)

colSelectServer(
  id,
  col_names,
  default_methods,
  col_selection = c("enable", "disable", "hide")
)

colSelectApp(
  col_names,
  default_methods,
  col_selection = "enable",
  testing = FALSE
)

methodSelectUI(id)

methodSelectServer(
  id,
  dat_obj,
  default_methods,
  col_selection = c("enable", "disable", "hide")
)

methodSelectApp(col_selection = "enable", testing = FALSE)
}
\arguments{
\item{id}{id to link shiny modules}

\item{col_names}{Column names.}

\item{default_methods}{Default methods.}

\item{col_selection}{string; one of 'enable', 'disable', or 'hide'. If 'enable', allow user to select which columns should be used for each variable. If 'disable', show columns used, but without the option to select. If 'hide', hide the column.}

\item{testing}{logical; passed to \code{shiny::shinyApp(..., options = list(test.mode))}}
}
\value{
A shiny app.

A shiny app.
}
\description{
This shiny module gives the user the option to manually map column names to
NACC variable names.

A short description...

This shiny module gives the user the option to manually map column names to
NACC variable names.

A short description...
}
