% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source.R
\name{data_load}
\alias{data_load}
\title{Load data}
\usage{
data_load(source, ...)
}
\arguments{
\item{source}{The data source object for dispatch.}

\item{...}{Arguments passed to S7 methods.}
}
\value{
The loaded data.
}
\description{
A short description...
}
