% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source.R
\name{data_source}
\alias{data_source}
\title{data_source}
\usage{
data_source(name = character(0), id = character(0), package = character(0))
}
\arguments{
\item{name}{Character. Human-readable display name shown in the UI dropdown.}

\item{id}{Character. Unique identifier used internally as a key.}

\item{package}{Character. The name of the package defining this data source.}
}
\description{
S7 class used to create data sources
}
