% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source.R
\name{data_source_server}
\alias{data_source_server}
\title{data_source_server generic}
\usage{
data_source_server(source, ...)
}
\arguments{
\item{source}{An S7 object representing a data source.}

\item{...}{Arguments passed to S7 methods.}
}
\value{
An S7 server object.
}
\description{
A short description...
}
