% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source.R
\name{data_source_ui}
\alias{data_source_ui}
\title{Data source UI}
\usage{
data_source_ui(source, ...)
}
\arguments{
\item{source}{An S7 data source object.}

\item{...}{Arguments passed to S7 methods.}
}
\value{
A Shiny UI element.
}
\description{
A short description...
}
