% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSelectModule.R
\name{dataSelectServer}
\alias{dataSelectServer}
\title{Data Selection Server Module}
\usage{
dataSelectServer(id)
}
\arguments{
\item{id}{A single string representing the module's identifier.}
}
\value{
A \code{shiny::moduleServer} function that returns a list containing:
\itemize{
\item \code{dat_obj}: A reactive value holding the loaded data, which is expected
to be a \code{data_nacc} object.
\item \code{data_source_extras}: A reactive value holding a list of additional
data source parameters or values.
}
Errors will be raised if the loaded data is not a \code{data_nacc} object.
}
\description{
A short description...
}
