% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSelectModule.R
\name{dataSelectUI}
\alias{dataSelectUI}
\title{Data selection UI module}
\usage{
dataSelectUI(id)
}
\arguments{
\item{id}{A single string.}
}
\value{
A \code{bslib::card} containing UI elements for selecting a data source and related options.
}
\description{
A short description...
}
