% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSelectModule_v2.R
\name{dataSelectUI_v2}
\alias{dataSelectUI_v2}
\title{Data selection UI module}
\usage{
dataSelectUI_v2(id)
}
\arguments{
\item{id}{A single string.}
}
\value{
A \code{bslib::card} containing UI elements for selecting a data source and related options.
}
\description{
A short description...
}
