% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_data_sources.R
\name{discover_data_sources}
\alias{discover_data_sources}
\title{Discover all available data source extensions}
\usage{
discover_data_sources()
}
\value{
A named list of \code{data_source} objects, keyed by their \code{id}.
}
\description{
Finds built-in data sources (defined in this package) and data sources from extensions
(from installed packages that declare themselves via DESCRIPTION fields).
}
