% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_gtExtras_for_my_gt_blt_bar_pct.R
\name{rgba_to_hex}
\alias{rgba_to_hex}
\title{Helpers from gtExtras used in \code{my_gt_blt_bar_pct}}
\usage{
rgba_to_hex(colors)
}
\description{
Helpers from gtExtras used in \code{my_gt_blt_bar_pct}
}
\keyword{internal}
