% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_table.R
\name{ht_cell_fill}
\alias{ht_cell_fill}
\title{Build a fill-style (background-color) spec for \code{\link[=ht_tab_style]{ht_tab_style()}}.}
\usage{
ht_cell_fill(color)
}
\arguments{
\item{color}{Background color.}
}
\description{
Build a fill-style (background-color) spec for \code{\link[=ht_tab_style]{ht_tab_style()}}.
}
