% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_table.R
\name{ht_cols_hide}
\alias{ht_cols_hide}
\title{Hide one or more columns.}
\usage{
ht_cols_hide(x, columns)
}
\arguments{
\item{x}{A \code{html_table}.}

\item{columns}{Character vector of column names.}
}
\description{
Hide one or more columns.
}
