% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_table.R
\name{ht_cols_label}
\alias{ht_cols_label}
\title{Relabel one or more columns.}
\usage{
ht_cols_label(x, ...)
}
\arguments{
\item{x}{A \code{html_table}.}

\item{...}{Named arguments mapping \code{column_name = label}. Labels may be
strings or shiny tags (incl. \code{shiny::HTML()}).}
}
\description{
Relabel one or more columns.
}
