% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_table.R
\name{ht_render}
\alias{ht_render}
\title{Render a \code{html_table} to \code{shiny::tags}}
\usage{
ht_render(x)
}
\arguments{
\item{x}{A \code{html_table}.}
}
\value{
A \code{shiny.tag} (the \verb{<table>} element).
}
\description{
Wrap the final builder result with \code{ht_render()} before passing it to
\code{shiny::renderUI()}:
}
\details{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{output$tbl <- shiny::renderUI(ht_render(my_table_fn(...)))
}\if{html}{\out{</div>}}
}
