% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_table.R
\name{ht_tab_spanner}
\alias{ht_tab_spanner}
\title{Add a column spanner across one or more columns.}
\usage{
ht_tab_spanner(x, label, columns)
}
\arguments{
\item{x}{A \code{html_table}.}

\item{label}{String or tag rendered above the spanned columns.}

\item{columns}{Character vector of column names to span.}
}
\description{
Add a column spanner across one or more columns.
}
