% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_table.R
\name{ht_tab_stubhead}
\alias{ht_tab_stubhead}
\title{Set the stubhead label (top-left cell, only shown when \code{rowname_col} is set).}
\usage{
ht_tab_stubhead(x, label)
}
\arguments{
\item{x}{A \code{html_table}.}

\item{label}{String or tag.}
}
\description{
Set the stubhead label (top-left cell, only shown when \code{rowname_col} is set).
}
