% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longTableModule.R
\name{longTableUI}
\alias{longTableUI}
\alias{longTableServer}
\alias{longTableApp}
\title{Shiny Module to Display Longitudinal Trends in Table}
\usage{
longTableUI(id)

longTableServer(
  id,
  dat,
  descriptions = c(Impaired = 0.03, Borderline = 0.1, `Low Average` = 0.26, Average =
    0.76, `High Average` = 0.92, Superior = 0.97, `Very Superior` = 1),
  fill_values = NULL,
  methods = "infer",
  table_font_size = 100,
  print_updating = F
)

longTableApp(
  dat,
  descriptions = c(Impaired = 0.03, Borderline = 0.1, `Low Average` = 0.26, Average =
    0.76, `High Average` = 0.92, Superior = 0.97, `Very Superior` = 1),
  fill_values = NULL,
  methods = "infer",
  table_font_size = 100,
  print_updating = F
)
}
\arguments{
\item{id}{string to tie UI to Server}

\item{dat}{A data.table object.}

\item{descriptions}{A named numeric vector. Default has names \code{"Impaired"}, \code{"Borderline"}, \code{"Low Average"}, \code{"Average"}, \code{"High Average"}, \code{"Superior"}, and \code{"Very Superior"} with corresponding values \code{0.03}, \code{0.10}, \code{0.26}, \code{0.76}, \code{0.92}, \code{0.97}, and \code{1}.}

\item{fill_values}{Optional.}

\item{methods}{Either a list containing methods used for standardization (each a character vector with named entried \code{method} and \code{version}), or \code{"infer"} (default). If \code{"infer"}, then methods are pulled from attributes of standardized columns.}

\item{table_font_size}{A numeric value passed to \code{gt::tab_options(table.font.size = gt::pct(table_font_size))}. Defaults to 100.}

\item{print_updating}{logical (default \code{FALSE}); should message be printed to let user know the table is updating. For debugging.}
}
\value{
A shiny app.
}
\description{
Table version of the plotCogVarModule.

A short description...
}
