% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainTableModule.R
\name{mainTableUI}
\alias{mainTableUI}
\alias{mainTableServer}
\alias{mainTableApp}
\title{Shiny Module to Display NACC T-Cog Neuropsychological Assessment Summary Table}
\usage{
mainTableUI(id)

mainTableServer(
  id,
  dat,
  descriptions = c(Impaired = 0.03, Borderline = 0.1, `Low Average` = 0.26, Average =
    0.76, `High Average` = 0.92, Superior = 0.97, `Very Superior` = 1),
  fill_values = NULL,
  table_font_size = 100,
  include_caption = F,
  print_updating = F
)

mainTableApp(dat, testing = FALSE)
}
\arguments{
\item{id}{An ID string to match module UI and server UI}

\item{dat}{Data to use. Must have exactly one row (data from one participant and one visit).}

\item{descriptions}{A named vector giving cutoffs to use for creating descriptions.}

\item{fill_values}{(optional) A named vector of same length (with same names)
as \code{descriptions} with hex color values to use. If \code{NULL}, no colorcoding used.
By default, evenly spread out colors from red through yellow to green are used.}

\item{table_font_size}{(optional) scalar indicating font size as a percentage.}

\item{include_caption}{Logical; should a caption be included above the table
with age and IQCODEs at current visit?}

\item{print_updating}{logical (defualt \code{FALSE}); should a message be displayed when table is being updated? For debugging.}

\item{testing}{logical (default \code{FALSE}); should the app be run in testing mode?}
}
\value{
a Shiny app object.
}
\description{
Shiny Module to Display NACC T-Cog Neuropsychological Assessment Summary Table
}
