% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainTableModule.R
\name{make_pdf_filename}
\alias{make_pdf_filename}
\title{Generate PDF filename for download}
\usage{
make_pdf_filename(naccid, visitdate, time = Sys.time())
}
\arguments{
\item{naccid}{Character string with the NACC ID.}

\item{visitdate}{Character string or Date with the visit date.}

\item{time}{POSIXct timestamp used for the "created-on" portion of the
filename. Defaults to \code{\link[=Sys.time]{Sys.time()}}.}
}
\value{
A character string suitable for use as a PDF filename.
}
\description{
Generate PDF filename for download
}
\keyword{internal}
