% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainTableModule.R
\name{make_table_caption}
\alias{make_table_caption}
\title{Generate table caption from ID and visit date}
\usage{
make_table_caption(naccid, visitdate)
}
\arguments{
\item{naccid}{Character vector of NACC IDs (will be uniquified).}

\item{visitdate}{Character or Date vector of visit dates (will be uniquified).}
}
\value{
A single character string caption.
}
\description{
Generate table caption from ID and visit date
}
\keyword{internal}
