% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodSelectModule.R
\name{methodSelectUI}
\alias{methodSelectUI}
\alias{methodSelectServer}
\alias{methodSelectApp}
\title{Shiny Module for Selecting Domains and Standardization Methods}
\usage{
methodSelectUI(id)

methodSelectServer(id, dat_obj, default_methods)

methodSelectApp(testing = FALSE)
}
\arguments{
\item{id}{id to link shiny modules}

\item{dat_obj}{reactive data object containing the loaded data.}

\item{default_methods}{named list. Each entry should be named after a
variable. The entry should be a named character vector with two elements:
one named 'method' to indicate the standardization method to use, and one
named 'version' to indicate the version to use for standardization.}

\item{testing}{logical; passed to \code{shiny::shinyApp(..., options = list(test.mode))}}
}
\value{
A shiny app.
}
\description{
This shiny module displays npsych_scores variables from the loaded data
and lets the user assign cognitive domains and standardization methods
using dropdown menus in an HTML table.
}
