% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nacc_groups.R
\docType{data}
\name{nacc_groups}
\alias{nacc_groups}
\title{NACC Variable Groups}
\format{
A named character vector where names are NACC variables, and entries are
corresponding groups.
}
\usage{
nacc_groups
}
\description{
A character vector giving groupings/domains of NACC scores. Meant to be
to ensure correct ordering.
}
\keyword{datasets}
