% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source.R
\name{new_data_source}
\alias{new_data_source}
\title{Create a new data source class}
\usage{
new_data_source(name, id, package, ...)
}
\arguments{
\item{name}{Character. Human-readable display name shown in the UI dropdown.}

\item{id}{Character. Unique identifier used internally as a key.}

\item{package}{Character. The name of the package defining this data source.}

\item{...}{Additional arguments passed to \code{\link[=data_source]{data_source()}}.}
}
\value{
An S7 class that inherits from \code{data_source}.
}
\description{
Convenience constructor that creates an S7 class inheriting from
\link{data_source}. The returned object is a \strong{class} (not an instance) that
can be instantiated with no arguments.
}
\examples{
my_source <- new_data_source(name = "My Source", id = "my_source", package = "ntrd")
my_source()

}
