% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntrd-package.R
\docType{package}
\name{ntrd-package}
\alias{ntrd}
\alias{ntrd-package}
\title{ntrd: Neuropsychological Test Results Dashboard}
\description{
Building on the ntrs package, this package includes a Shiny application that makes it easy to present test scores in a way useful to consensus diagnosis. Data can be imported as a .csv file, or extensions can be created to import directly from REDCap/any other database system accessible via API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rmtrane.github.io/ntrd}
  \item Report bugs at \url{https://github.com/rmtrane/ntrd/issues}
}

}
\author{
\strong{Maintainer}: Ralph Møller Trane \email{rtrane@wisc.edu} (\href{https://orcid.org/0000-0001-7824-1785}{ORCID})

}
\keyword{internal}
