% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModule.R
\name{plotUI}
\alias{plotUI}
\alias{plotServer}
\alias{plotApp}
\title{Plot Module}
\usage{
plotUI(id)

plotServer(
  id = "Attention/Processing",
  dat,
  x_range = c("2016-05-01", "2020-08-01"),
  y_range = c(-2.5, 2.5),
  descriptions = c(Impaired = 0.03, Borderline = 0.1, `Low Average` = 0.26, Average =
    0.76, `High Average` = 0.92, Superior = 0.97, `Very Superior` = 1),
  fill_values = calc_fill_colors(n = 7),
  print_updating = T,
  shade_descriptions = TRUE,
  new_id = NULL,
  var_groups = nacc_var_groups
)

plotApp(dat_input = prepare_data(demo_data), studyids = NULL, testing = FALSE)
}
\arguments{
\item{id}{A single string. Optional, defaults to \code{"Attention/Processing"}. Should be one of \code{nacc_var_groups}.}

\item{dat}{A data.table. Should be participant specific.}

\item{x_range}{A date range as a character vector with two elements.}

\item{y_range}{A numeric vector with two elements. Optional, defaults to \code{c(-2.5, 2.5)}.}

\item{descriptions}{A named numeric vector. Optional, defaults to the vector described in the function definition.}

\item{fill_values}{A vector of fill colors. Optional, defaults to \code{calc_fill_colors(n = 7)}.}

\item{print_updating}{A logical value. Optional, defaults to \code{TRUE}.}

\item{shade_descriptions}{A logical value indicating if the plots should be shaded according to the regions given by \code{descriptions} with colors given by \code{fill_values}. Optional, defaults to \code{TRUE}.}

\item{new_id}{Optional. String to be used for table ID. If \code{NULL} (default), random string assigned.}

\item{var_groups}{A named character vector mapping variable names to domain groups. Can be a reactive. Defaults to \code{nacc_var_groups}.}

\item{dat_input}{A data frame. Defaults to \code{prepare_data(demo_data)}.}

\item{studyids}{Optional. If \code{NULL} (default), will use unique values from \code{dat_input$NACCID}.}

\item{testing}{Logical, whether to run the app in testing mode.}
}
\value{
A \code{shiny::tagList} object containing two \code{shiny::conditionalPanel}'s:
one with the plot, and one with text for when no standardized scores
were found. Both wrapped in a \code{bslib::accordion_panel} for inclusion in
main app.

No return value.

A shiny app.
}
\description{
UI for displaying plots of standardized scores across visits.

Server logic to handle the creation and updating of plots to show standardized scores across visits.

Shiny app using the plotUI and plotServer.
}
