% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyAddResources.R
\name{shinyAddResources}
\alias{shinyAddResources}
\title{Add Shiny resources}
\usage{
shinyAddResources(development)
}
\arguments{
\item{development}{logical; indicading if run in development mode. If \code{TRUE}, resources will be pulled from \code{inst/www} and \code{inst/qmd}. Otherwise, the path for the installed package will be used.}
}
\value{
No return value. This function is called for its side effects of adding
resource paths for Shiny applications, mapping \code{"www"} and \code{"qmd"}
to their respective directories within the package installation or
development environment.
}
\description{
Adds resource paths for Shiny applications to access static files.
}
