% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_header.R
\name{shinyApp_header}
\alias{shinyApp_header}
\title{Shiny app header}
\usage{
shinyApp_header()
}
\value{
A \code{shiny::tagList} used to add \code{www/scripts.js} and \code{www/styles.css} to the head
of the app, and adding the spinner layers.
}
\description{
Shiny app header
}
