% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyAssessmentApp_v2.R
\name{shinyAssessmentApp_v2}
\alias{shinyAssessmentApp_v2}
\title{Run Shiny App}
\usage{
shinyAssessmentApp_v2(testing = FALSE)
}
\arguments{
\item{testing}{Logical: is the app run for testing purposes?}
}
\value{
Runs the Shiny app.
}
\description{
Wrapper that runs the Shiny application.
}
