% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyDashboard.R
\name{shinyDashboard}
\alias{shinyDashboard}
\title{Run Shiny App}
\usage{
shinyDashboard(testing = FALSE)
}
\arguments{
\item{testing}{Logical: is the app run for testing purposes?}
}
\value{
Runs the Shiny app.
}
\description{
Wrapper that runs the Shiny application.
}
