% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension_updates.R
\name{try_update}
\alias{try_update}
\title{Try update extension}
\usage{
try_update(pkg)
}
\arguments{
\item{pkg}{A single string.}
}
\value{
\code{TRUE} if the update executes successfully, \code{FALSE} otherwise. A message
is emitted if an error occurs.
}
\description{
Try update extension
}
