% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wadrc_source.R
\name{wadrc_source}
\alias{wadrc_source}
\title{Wisconsin ADRC data source}
\usage{
wadrc_source()
}
\arguments{
\item{ns}{A Shiny namespace function.}

\item{id}{A single string, the Shiny module ID.}

\item{source}{An object of class \code{wadrc_source}.}

\item{uds2_api_token}{A single string.}

\item{uds3_api_token}{A single string.}

\item{uds4_api_token}{A single string.}

\item{...}{Currently unused; must be empty.}
}
\value{
An S7 object of class \code{wadrc_source}, inheriting from \code{data_source}, with
\code{name = "Wisconsin ADRC"} and \code{id = "wadrc_redcap"}.

A \code{shiny::tagList} containing UI elements for configuring REDCap API tokens and a Panda API Key.

A \code{shiny::moduleServer} function that returns a list containing:
\itemize{
\item \code{params}: A reactive list of API tokens, requiring \code{uds2_api_token},
\code{uds3_api_token}, and \code{uds4_api_token} to be present.
\item \code{restore}: A function to update the API token input fields.
\item \code{extras}: A \code{shiny::reactiveValues} object holding \code{all_values} (data
fetched using \code{panda_api_token}) and \code{panda_api_token}.
\item \code{session}: The Shiny session object.
}

An object of class \code{data_nacc} if data is successfully retrieved and combined.
If no data is retrieved from any UDS, \code{NULL} is returned and an error notification is shown in a Shiny context.
Notifications are displayed within a Shiny application to indicate data loading progress and completion.
}
\description{
A short description...

A short description...
}
\keyword{internal}
