% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_source.R
\name{data_source}
\alias{data_source}
\title{data_source}
\usage{
data_source(name = character(0), id = character(0))
}
\arguments{
\item{name}{A character string for the display name.}

\item{id}{A character string for the unique identifier.}
}
\description{
A short description...
}
