% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodSelectModule.R
\name{methodSelectUI}
\alias{methodSelectUI}
\alias{methodSelectServer}
\alias{methodSelectApp}
\title{Shiny Module for Selecting Columns}
\usage{
methodSelectUI(id)

methodSelectServer(
  id,
  dat_obj,
  default_methods,
  col_selection = c("enable", "disable", "hide")
)

methodSelectApp(col_selection = "enable", testing = FALSE)
}
\arguments{
\item{id}{id to link shiny modules}

\item{default_methods}{named list. Each entry should be named after a variable. The entry should be a named character vector with to elements: one named 'method' to indicate the standardization method to use, and one name 'version' to indicate the version to use for standardization.}

\item{col_selection}{string; one of 'enable', 'disable', or 'hide'. If 'enable', allow user to select which columns should be used for each variable. If 'disable', show columns used, but without the option to select. If 'hide', hide the column.}

\item{testing}{logical; passed to \code{shiny::shinyApp(..., options = list(test.mode))}}

\item{col_names}{reactive character vector with names to choose from}
}
\value{
A shiny app.
}
\description{
This shiny module gives the user the option to manually map column names to
NACC variable names.

A short description...
}
